# New implementation for writing mat result files