--- a/SimulationRuntime/c/simulation/results/CMakeLists.txt
+++ b/SimulationRuntime/c/simulation/results/CMakeLists.txt
@@ -7,7 +7,7 @@ simulation_result.cpp      simulation_result_ia.cpp   simulation_result_plt.cpp
 simulation_result_csv.cpp  simulation_result_mat.cpp  simulation_result_wall.cpp
 )
 
-SET(results_headers ../../util/read_csv.h 
+SET(results_headers ../../util/read_csv.h
 simulation_result.h      simulation_result_ia.h   simulation_result_plt.h
 simulation_result_csv.h  simulation_result_mat.h  simulation_result_wall.h
 )