--- a/SimulationRuntime/c/simulation/results/CMakeLists.txt
+++ b/SimulationRuntime/c/simulation/results/CMakeLists.txt
@@ -4,12 +4,12 @@
 # Quellen und Header
 SET(results_sources
 simulation_result.cpp      simulation_result_ia.cpp   simulation_result_plt.cpp
-simulation_result_csv.cpp  simulation_result_mat.cpp  simulation_result_wall.cpp
+simulation_result_csv.cpp  simulation_result_mat4.cpp  simulation_result_wall.cpp    MatVer4.cpp
 )
 
 SET(results_headers ../../util/read_csv.h
 simulation_result.h      simulation_result_ia.h   simulation_result_plt.h
-simulation_result_csv.h  simulation_result_mat.h  simulation_result_wall.h
+simulation_result_csv.h  simulation_result_mat4.h  simulation_result_wall.h  MatVer4.h
 )
 
 # Library util