--- a/SimulationRuntime/c/simulation/simulation_runtime.cpp
+++ b/SimulationRuntime/c/simulation/simulation_runtime.cpp
@@ -71,6 +71,7 @@
 #include "simulation/results/simulation_result_plt.h"
 #include "simulation/results/simulation_result_csv.h"
 #include "simulation/results/simulation_result_mat.h"
+#include "simulation/results/simulation_result_mat4.h"
 #include "simulation/results/simulation_result_wall.h"
 #include "simulation/results/simulation_result_ia.h"
 #include "simulation/solver/solver_main.h"
@@ -548,11 +549,17 @@ int initializeResultData(DATA* simData, threadData_t *threadData, int cpuTime)
     sim_result.emit = omc_csv_emit;
     /* sim_result.writeParameterData = omc_csv_writeParameterData; */
     sim_result.free = omc_csv_free;
-  } else if(0 == strcmp("mat", simData->simulationInfo->outputFormat)) {
+  } else if(0 == strcmp("mat-old", simData->simulationInfo->outputFormat)) {
     sim_result.init = mat4_init;
     sim_result.emit = mat4_emit;
     sim_result.writeParameterData = mat4_writeParameterData;
     sim_result.free = mat4_free;
+    resultFormatHasCheapAliasesAndParameters = 1;
+  } else if(0 == strcmp("mat", simData->simulationInfo->outputFormat)) {
+    sim_result.init = mat4_init4;
+    sim_result.emit = mat4_emit4;
+    sim_result.writeParameterData = mat4_writeParameterData4;
+    sim_result.free = mat4_free4;
     resultFormatHasCheapAliasesAndParameters = 1;
 #if !defined(OMC_MINIMAL_RUNTIME)
   } else if(0 == strcmp("wall", simData->simulationInfo->outputFormat)) {