**Changes**
- [x] Reduce file size for constant variables
- [x] Full support for single precision (parameters & variables)
- [x] Improve performance
- [x] Add unit information

**Tickets**
* [ticket:4574](https://trac.openmodelica.org/OpenModelica/ticket/4574)
* [ticket:4582](https://trac.openmodelica.org/OpenModelica/ticket/4582)