@OpenModelica-Hudson 2017-10-27
The job failed to compile; for details, see [OpenModelica_TEST_PULL_REQUEST 2017-10-27_21-24-28](https://test.openmodelica.org/hudson//job/OpenModelica_TEST_PULL_REQUEST/4825/console).