@OpenModelica-Hudson 2017-10-27
The test suite is unstable according to [OpenModelica_TEST_PULL_REQUEST 2017-10-27_21-35-20](https://test.openmodelica.org/hudson//job/OpenModelica_TEST_PULL_REQUEST/4827/).