@OpenModelica-Hudson 2017-10-27
The tests run correctly according to [OpenModelica_TEST_PULL_REQUEST 2017-10-28_00-30-29](https://test.openmodelica.org/hudson//job/OpenModelica_TEST_PULL_REQUEST/4830/console).