@lochel 2017-10-28
I did a quick test running `Modelica.Mechanics.MultiBody.Examples.Loops.EngineV6_analytic` with the old and new mat format. I did the timings a couple of times, always with similar results. However, I don't trust them too much. But at least the file size went down significant.

|   | old | new | ratio |
|---|---|---|---|
| time | 0.309584s | 0.159811s |  0,52% | 
| file size | 21574936B | 14109074B | 65% |