@sjoelund 2017-10-28
The time required to write the file should be roughly linear to the file size. Do you still look for the filtered flag in each iteration like the old mat output or is it done only once (saving indexes somewhere and looping only over that array?); I guess it wouldn't help terribly much though