@OpenModelica-Hudson 2017-10-28
The tests run correctly according to [OpenModelica_TEST_PULL_REQUEST 2017-10-28_14-25-25](https://test.openmodelica.org/hudson//job/OpenModelica_TEST_PULL_REQUEST/4831/console).