@OpenModelica-Hudson 2017-10-30
The tests run correctly according to [OpenModelica_TEST_PULL_REQUEST 2017-10-30_08-57-12](https://test.openmodelica.org/hudson//job/OpenModelica_TEST_PULL_REQUEST/4833/console).